Fix UI alignment issues and out of bound index access
When the configuration selector was not returning with
a new description of the added VM, a possible out
of bounds access was performed because we were adding
at a wrong time the NO_VM_CONFIGS element, right after
the start of the activity. Move the check of the list
after the activity returns from the VmActivity to
prevent adding NO_VM_CONFIGS when the list is not empty.

Fix a bunch of layout issues and move everything from
LinearLayout to ConstraintLayout.

Signed-off-by: Seb <seb@Sebs-MacBook-Pro.local>